add clone/cd instructions for building the book
As of now the [Crystal docs](https://crystal-lang.org/docs/index.html) make no reference of this repo. It took me a bit of googling to figure out what I was supposed to build the book from. Adding instructions that point back to this repo would've been helpful, so here's a PR :smile: