Update http_server.md (#256)
* Update http_server.md

Hello, I noticed that the current tutorial isn't updated to the current version and there was a break change with the port setting which is now done separately. 

I am not sure how good it is but I hope it helps.

* Fixed a few mistakes and added explenation

Hi, 
I changed a few thing I did wrong over there and added explanation to the new line that is responsible to setting the port of the HTTP Server. 

If you have feedback I'd love to hear it, so I could get better next time.