Replace inner links: .html -> .md
This is needed for migration to MkDocs, but GitBook also translates the links seamlessly.
It's an improvement on its own because now the links are also valid in the Markdown world (e.g. browsing on GitHub).

This was done by replacing the regex /([( ][^\]:` ]+)\.html/ with "\1.md"
with a further check that "index.md" shouldn't be a thing