Crystal Governance document (#512)
The Crystal Governance document. This document is based on the existing experience in Python, and is fairly minimal on purpose. The immediate goal is to have something on which to based the key decision-makings processes for Crystal, and is one of the on-going efforts to improve the communication between the different boards and the community.